producer/requirements.txt
requests==2.32.0
kafka-python==2.0.2